refactor UserPosts as a page and extends Posts component
