+++ src/components/Blog/Posts/index.js
@@ -1,26 +1,16 @@
 import { React, useEffect, useState } from "react";
-import { useSelector } from 'react-redux';
 import { Link } from "react-router-dom";
-import axios from "axios";
+import moment from "moment";
 import classes from "./index.module.css";
 
-function Posts() {
-    const auth = useSelector((state) => state.auth.value);
+function Posts(props) {
     const [isLoading, setIsLoading] = useState(true);
     const [posts, setPosts] = useState([]);
 
     useEffect(() => {
-        axios.get('http://localhost:3001/api/blog/posts', {
-            headers: { Authorization: `Bearer ${auth.bearerToken}` }
-        })
-            .then((response) => {
-                setIsLoading(false);
-                setPosts(response.data.data);
-            })
-            .catch((error) => {
-                console.log(error);
-            });
-    }, [auth.bearerToken])
+        setIsLoading(false);
+        setPosts(props.posts);
+    }, [props.posts]);
 
     if (isLoading) {
         return (
@@ -37,7 +27,7 @@ function Posts() {
                     <div className={classes.postHeader}>
                         <Link to={`/blog/posts/${post.id}/${post.slug}`}>
                             <span className={classes.left}>{post.title}</span>
-                            <span className={classes.right}>{post.date}</span>
+                            <span className={classes.right}>{moment(post.date).format("MMMM D, YYYY")}</span>
                         </Link>
                     </div>
                     <div className={classes.postContent}>
