+++ src/components/Blog/Posts/index.module.css
@@ -1,8 +1,5 @@
 .main {
     width: 100%;
-    float: left;
-    overflow: auto;
-    clear: both;
 }
 
 .post {
