--- src/components/Blog/UserPosts/index.module.css
@@ -1,28 +1,0 @@
-.main {
-    width: 100%;
-}
-
-.post {
-    background-color: blueviolet;
-    margin-bottom: 32px;
-}
-
-.postHeader {
-    padding: 1rem 0.4rem;
-    background-color: green;
-    font-size: 1.2rem;
-}
-
-.postHeader a {
-    text-decoration: none;
-}
-
-.postContent {
-    padding: 0.4rem;
-    text-indent: 1rem;
-}
-
-.right {
-    float: right;
-    overflow: auto;
-}
